Fri, Jun 10, 2022, 7:46 PM +00:00

+++ home/coins.c
@@ -3,7 +3,7 @@
 
 
 float get_money();
-int troco(float v, float m);
+int troco(double v, double m);
 
 int main(void)
 {
